Add screen Execute Routine

+++ app/src/main/java/ar/edu/itba/hci/fitcenter/RoutineList.kt
@@ -40,7 +40,7 @@ import ar.edu.itba.hci.fitcenter.ui.theme.FitcenterTheme
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-fun formatDate(dateValue: Int): String {
+fun formatDate(dateValue: Long): String {
     val date = LocalDate.parse(dateValue.toString(), DateTimeFormatter.BASIC_ISO_DATE)
     val formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy") // You can adjust the pattern as needed
     return date.format(formatter)
